Business Logic Implementation && SQL Script Update

+++ app/tandem/src/main/java/com/tandem/repository/ContentRepository.java
@@ -8,12 +8,12 @@ import org.springframework.data.jpa.repository.Query;
 
 public interface ContentRepository extends JpaRepository<ContentEntity, Long> {
     @Procedure(name = "content_management.add_content")
-    void addContent(@Param("p_photo") Integer photoId, @Param("p_video") Integer videoId,
-                    @Param("p_audio") Integer audioId, @Param("p_text") Integer textId);
+    void addContent(@Param("p_photo") Long photoId, @Param("p_video") Long videoId,
+                    @Param("p_audio") Long audioId, @Param("p_text") Long textId);
 
     @Procedure(name = "content_management.delete_content")
-    void deleteContent(@Param("p_content_id") Integer contentId);
+    void deleteContent(@Param("p_content_id") Long contentId);
 
     @Query(value = "SELECT content_management.check_content_deletable(:p_content_id)", nativeQuery = true)
-    boolean checkContentDeletable(@Param("p_content_id") int contentId);
+    boolean checkContentDeletable(@Param("p_content_id") Long contentId);
 }
